+++ CHANGES.md
@@ -5,6 +5,7 @@ Change Log
     - Added example test data to docs/testdata directory
     - Added reprint button to Order print view
     - Updated lite badge price, day
+    - Kumo Lite 2018 badge format added
 
 - Version 3.0.4 (6/18/2018)
     - Show parental consent form checkbox in detail form